requirements.txt
mysql-connector-python>=8.0.33
python-dotenv>=1.0.0
msal>=1.24.0
requests>=2.31.0

